Support Windows 11

+++ AutoHotkey/C0ffee Anti Idle v2.ahk
@@ -91,14 +91,29 @@ ReloadScript()
 
 HighContrastOn()
 {
-	Run "`"C:\Windows\Resources\Ease of Access Themes\hc2.theme`""
-	sleep 5000
-	run "taskkill /im systemsettings.exe /f"
+	file := "C:\Windows\Resources\Ease of Access Themes\hc2.theme"
+	if FileExist(file) {
+		tooltip "grr"
+		Run "C:\Windows\Resources\Ease of Access Themes\hc2.theme"
+		sleep 1000
+		run "taskkill /im systemsettings.exe /f"
+	}
 }
 
 HighContrastOff()
 {
-	Run "C:\Windows\Resources\Themes\Light.theme"
-	sleep 5000
-	run "taskkill /im systemsettings.exe /f"
+	file := "C:\Windows\Resources\Themes\Light.theme"
+	if FileExist(file) {
+		Run "C:\Windows\Resources\Themes\Light.theme"
+		sleep 1000
+		run "taskkill /im systemsettings.exe /f"
+	} 
+	
+	file := "C:\Windows\Resources\Themes\aero.theme"
+	if FileExist(file) {
+		Run "C:\Windows\Resources\Themes\aero.theme"
+		sleep 1000
+		run "taskkill /im systemsettings.exe /f"
+	}
+	
 }
